+++ kda-bot/test/src/me/ricky/kda/bot/test/Main.kt
@@ -7,14 +7,13 @@ import club.minnced.jda.reactor.toMono
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.JsonElement
+import kotlinx.serialization.serializer
 import me.ricky.kda.core.KDA
-import me.ricky.kda.core.command.Command
-import me.ricky.kda.core.command.CommandContext
-import me.ricky.kda.core.command.CommandManager
-import me.ricky.kda.core.command.CommandManagerTranslation
+import me.ricky.kda.core.command.*
 import me.ricky.kda.core.lang.Text
 import me.ricky.kda.core.lang.loadTranslations
 import net.dv8tion.jda.api.events.message.guild.GuildMessageReceivedEvent
+import java.util.*
 
 @Serializable
 data class Translation(
@@ -29,11 +28,13 @@ object TestCommand : Command<TestCommand.Translation> {
 
   override fun CommandContext<Translation>.execute() {
     channel.sendMessage(translation.message("{guild}" to guild.name)).queue()
+    channel.sendMessage(translation.otherMessage).queue()
   }
 
   @Serializable
   data class Translation(
-    val message: Text
+    val message: Text,
+    val otherMessage: String,
   )
 }
 
@@ -43,7 +44,7 @@ fun main() {
     ?: error("Cannot start bot because no token was provided.")
 
   KDA(token) {
-    val commandManager = CommandManager(translation)
+    val commandManager = CommandManager(translation, ConstantServerDataManager(locale = Locale.CANADA))
     val eventManager = ReactiveEventManager()
 
     commandManager += TestCommand