+++ CHANGELOG.md
@@ -9,3 +9,6 @@
 * Added Command Manager
 * Redesigned Language System
 * Added Server Data Manager
+### 1.2.1
+* Translations will now get from default locale
+if there is no translation for the given locale.